# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ClaimedReportJobDto": {
        "properties": {
          "attempt": {
            "type": "number"
          },
          "leaseExpiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "leaseToken": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "reportId": {
            "format": "uuid",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotDto"
          }
        },
        "required": [
          "reportId",
          "leaseToken",
          "leaseExpiresAt",
          "attempt",
          "model",
          "promptVersion",
          "snapshot"
        ],
        "type": "object"
      },
      "CompleteReportJobDto": {
        "properties": {
          "generatedBy": {
            "enum": [
              "GEMINI",
              "TEMPLATE"
            ],
            "type": "string"
          },
          "leaseToken": {
            "type": "string"
          },
          "model": {
            "example": "gemini-3.5-flash-lite",
            "nullable": true,
            "type": "string"
          },
          "narrative": {
            "example": {
              "bullets": [
                "...",
                "...",
                "..."
              ],
              "headline": "ภาพรวมพนักงาน"
            },
            "type": "object"
          },
          "promptVersion": {
            "example": "employee-summary-v1",
            "type": "string"
          }
        },
        "required": [
          "leaseToken",
          "generatedBy",
          "model",
          "promptVersion",
          "narrative"
        ],
        "type": "object"
      },
...
      "EmptyBodyDto": {
        "properties": {},
        "type": "object"
      },
      "ErrorBodyDto": {
        "properties": {
          "code": {
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "currentReportId": {
            "description": "REPORT_IN_PROGRESS only",
            "format": "uuid",
            "nullable": true,
            "type": "object"
          },
          "currentVersion": {
            "description": "VERSION_CONFLICT only",
            "type": "number"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            },
            "type": "array"
          },
          "message": {
            "example": "Please correct the highlighted fields.",
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
      "ErrorEnvelopeDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "FailReportJobDto": {
        "properties": {
          "errorCode": {
            "enum": [
              "PROVIDER_TIMEOUT",
              "PROVIDER_RATE_LIMIT",
              "PROVIDER_UNAVAILABLE",
              "INVALID_MODEL_OUTPUT",
              "PROVIDER_AUTH_ERROR",
              "MODEL_UNAVAILABLE",
              "INVALID_SNAPSHOT",
              "CONTENT_REJECTED"
            ],
            "type": "string"
          },
          "leaseToken": {
            "type": "string"
          }
        },
        "required": [
          "leaseToken",
          "errorCode"
        ],
        "type": "object"
      },
      "FieldErrorDto": {
        "properties": {
          "code": {
            "example": "DECIMAL_SCALE_EXCEEDED",
            "type": "string"
          },
          "field": {
            "example": "salary",
            "type": "string"
          },
          "message": {
            "example": "Salary must have at most 2 decimal places.",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "JobStatusDto": {
        "properties": {
          "reportId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "type": "string"
          }
        },
        "required": [
          "reportId",
          "status"
        ],
        "type": "object"
      },
      "MetaDto": {
        "properties": {
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "type": "object"
      },
...
      "ReportSummaryDto": {
        "properties": {
          "completedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "errorCode": {
            "nullable": true,
            "type": "string"
          },
          "generatedBy": {
            "enum": [
              "GEMINI",
              "TEMPLATE"
            ],
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "model": {
            "nullable": true,
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "snapshotCapturedAt": {
            "format": "date-time",
            "type": "string"
          },
          "source": {
            "enum": [
              "MANUAL",
              "SCHEDULED"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "type": "string"
          },
          "totalEmployees": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "status",
          "source",
          "snapshotCapturedAt",
          "totalEmployees",
          "createdAt",
          "completedAt",
          "generatedBy",
          "model",
          "promptVersion",
          "errorCode"
        ],
        "type": "object"
      },
...
      "SnapshotDepartmentDto": {
        "properties": {
          "active": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "inactive": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "total",
          "active",
          "inactive"
        ],
        "type": "object"
      },
      "SnapshotDto": {
        "properties": {
          "activeEmployees": {
            "type": "number"
          },
          "businessDate": {
            "format": "date",
            "type": "string"
          },
          "capturedAt": {
            "format": "date-time",
            "type": "string"
          },
          "departments": {
            "items": {
              "$ref": "#/components/schemas/SnapshotDepartmentDto"
            },
            "type": "array"
          },
          "inactiveEmployees": {
            "type": "number"
          },
          "schemaVersion": {
            "enum": [
              1
            ],
            "type": "number"
          },
          "timezone": {
            "enum": [
              "Asia/Bangkok"
            ],
            "type": "string"
          },
          "totalEmployees": {
            "type": "number"
          }
        },
        "required": [
          "schemaVersion",
          "capturedAt",
          "businessDate",
          "timezone",
          "totalEmployees",
          "activeEmployees",
          "inactiveEmployees",
          "departments"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "cookie": {
        "in": "cookie",
        "name": "employee_console.local.sid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "NestJS API for Employee Console. Browser clients authenticate with the session cookie (Google OIDC) and send X-CSRF-Token on mutations. /internal/v1 is for the n8n worker/scheduler (bearer tokens).",
    "title": "Employee Console API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/internal/v1/report-jobs/claim": {
      "post": {
        "operationId": "InternalReports_claim",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyBodyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/ClaimedReportJobDto"
                        }
                      ],
                      "nullable": true
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "data is null when no job is due; updates the worker heartbeat",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR / INVALID_QUERY / INVALID_IDEMPOTENCY_KEY / MALFORMED_JSON",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "RATE_LIMITED / REPORT_QUOTA_EXCEEDED",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/v1/report-jobs/{id}/complete": {
      "post": {
        "operationId": "InternalReports_complete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteReportJobDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/JobStatusDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR / INVALID_QUERY / INVALID_IDEMPOTENCY_KEY / MALFORMED_JSON",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "EMPLOYEE_NOT_FOUND / REPORT_NOT_FOUND",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VERSION_CONFLICT / IDEMPOTENCY_CONFLICT / REPORT_IN_PROGRESS / STALE_LEASE / REPORT_ALREADY_COMPLETED",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "PAYLOAD_TOO_LARGE",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/v1/report-jobs/{id}/fail": {
      "post": {
        "operationId": "InternalReports_fail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FailReportJobDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/JobStatusDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "QUEUED (retry scheduled) or FAILED",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR / INVALID_QUERY / INVALID_IDEMPOTENCY_KEY / MALFORMED_JSON",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "EMPLOYEE_NOT_FOUND / REPORT_NOT_FOUND",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VERSION_CONFLICT / IDEMPOTENCY_CONFLICT / REPORT_IN_PROGRESS / STALE_LEASE / REPORT_ALREADY_COMPLETED",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/v1/reports/scheduled": {
      "post": {
        "operationId": "InternalReports_scheduled",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyBodyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReportSummaryDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Today's scheduled report already exists",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReportSummaryDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Created for today (Asia/Bangkok business day)",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR / INVALID_QUERY / INVALID_IDEMPOTENCY_KEY / MALFORMED_JSON",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VERSION_CONFLICT / IDEMPOTENCY_CONFLICT / REPORT_IN_PROGRESS / STALE_LEASE / REPORT_ALREADY_COMPLETED",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "AI_NOT_CONFIGURED / AUTH_NOT_CONFIGURED / DEPENDENCY_UNAVAILABLE",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "internal"
        ]
      }
    }
  },
  "servers": [],
...
}